+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Made `force update` link for a flow table
 ### Changed
 - Flow configuration separation
+### Fixed
+- Fixed error handling for the query explainer
 
 ## [0.33.0] - 2024-12-03
 ### Added
@@ -80,7 +82,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Lineage graph: redirect to another dataset
 - Application search: redirect to another dataset
 - Lineage tab displayed correctly
-- Prettified histopy timeline
+- Prettified history timeline
 
 ## [0.27.0] - 2024-09-25
 ### Added